requirements.txt
attrs==20.2.0
backcall==0.2.0
coverage==5.3
decorator==4.4.2
iniconfig==1.0.1
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
more-itertools==8.5.0
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.7
ptyprocess==0.6.0
py==1.9.0
Pygments==2.7.0
pyparsing==2.4.7
pytest==6.0.2
pytest-cov==2.10.1
python-decouple==3.3
selenium==3.141.0
six==1.15.0
toml==0.10.1
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
